feat: decimal coins, simplified elevation, and UI safety fixes
- Coins now use double with 2 decimal places throughout
- Database auto-migrates BIGINT to DOUBLE for coins columns
- Coin display uses human-readable suffixes (M, B, T, Qa, Qi, Sx, Sp, Oc, No, Dc)
- Payout calculated before multiplier increase (fixes earning order)
- Elevation simplified: value IS the multiplier directly (no level conversion)
- Removed "Lv." prefix from all elevation displays
- Added try-catch protection to page auto-refresh to prevent disconnects

+++ hyvexa-parkour-ascend/src/main/java/io/hyvexa/ascend/ascension/AscensionManager.java
@@ -31,7 +31,7 @@ public class AscensionManager {
      * Checks if a player can perform an Ascension.
      */
     public boolean canAscend(UUID playerId) {
-        long coins = playerStore.getCoins(playerId);
+        double coins = playerStore.getCoins(playerId);
         return coins >= AscendConstants.ASCENSION_COIN_THRESHOLD;
     }
 
@@ -41,7 +41,7 @@ public class AscensionManager {
      * @return the new Ascension count, or -1 if insufficient coins
      */
     public int performAscension(UUID playerId) {
-        long coins = playerStore.getCoins(playerId);
+        double coins = playerStore.getCoins(playerId);
         if (coins < AscendConstants.ASCENSION_COIN_THRESHOLD) {
             LOGGER.atInfo().log("[Ascension] Player " + playerId + " has insufficient coins: " + coins);
             return -1;
@@ -66,7 +66,7 @@ public class AscensionManager {
         }
 
         // Reset progress
-        progress.setCoins(0L);
+        progress.setCoins(0.0);
         progress.setElevationMultiplier(1);
         progress.clearSummitLevels();
 
@@ -84,7 +84,7 @@ public class AscensionManager {
 
         // Apply starting coins skill
         if (progress.hasSkillNode(SkillTreeNode.COIN_T1_STARTING_COINS)) {
-            progress.setCoins(1000L);
+            progress.setCoins(1000.0);
         }
 
         playerStore.markDirty(playerId);
